- Add second SPA for user (separate admin spa and user spa) - Make sanctum auth middlewar working for laravel routes (blades), login into SPA, session save in cookies, sanctum middlewar works for routes -

+++ app/Modules/Auth/Controllers/LoginController.php
@@ -33,12 +33,42 @@ class LoginController extends Controller
         $this->token = $token;
     }
 
-    /**
-     * Attempt to log the user into the application.
-     *
-     * @param  Request  $request
-     * @return bool
-     */
+     /**
+      * Attempt to log the user into the application.
+      *
+      * @param  Request  $request
+      * @return bool
+      */
+     public function login(Request $request)
+     {
+         $this->validateLogin($request);
+
+         // If the class is using the ThrottlesLogins trait, we can automatically throttle
+         // the login attempts for this application. We'll key this by the username and
+         // the IP address of the client making these requests into this application.
+         if (method_exists($this, 'hasTooManyLoginAttempts') &&
+             $this->hasTooManyLoginAttempts($request)) {
+             $this->fireLockoutEvent($request);
+
+             return $this->sendLockoutResponse($request);
+         }
+
+         if ($this->attemptLogin($request)) {
+             if ($request->hasSession()) {
+                 $request->session()->put('auth.password_confirmed_at', time());
+             }
+
+             return $this->sendLoginResponse($request);
+         }
+
+         // If the login attempt was unsuccessful we will increment the number of attempts
+         // to login and redirect the user back to the login form. Of course, when this
+         // user surpasses their maximum number of attempts they will get locked out.
+         $this->incrementLoginAttempts($request);
+
+         return $this->sendFailedLoginResponse($request);
+     }
+
     protected function attemptLogin(Request $request)
     {
         $user = User::where('email', strtolower($request->input($this->username())))->first();
@@ -55,7 +85,15 @@ class LoginController extends Controller
 
         $this->setToken($user->createToken($request->device_name)->plainTextToken);
 
-        return true;
+        if (method_exists($this->guard(), 'attempt')) {
+            $resp = $this->guard()->attempt(
+                $this->credentials($request), $request->filled('remember')
+            );
+        } else {
+            $resp = true;
+        }
+
+        return $resp;
     }
 
     /**
@@ -102,9 +140,22 @@ class LoginController extends Controller
      */
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
+        $session = auth()->guard('web')->getsession();
+        $cookieName = $session->getName();
+
+        if (method_exists($request->user()->currentAccessToken(), 'delete')) {
+            $request->user()->currentAccessToken()->delete();
+        } else {
+
+        }
+
         app()->get('auth')->forgetGuards();
         auth('web')->logout();
+        auth()->guard('web')->logout();
+
+        $delete_cookie = \Cookie::forget($cookieName);  // TODO: wrong, think about another way
+
+        return response()->noContent()->withCookie($delete_cookie);
     }
 
     /**
